feat(examples): configure trace density patterns
